Small cleanup of shaders
